Restructure actions into separate folder
