+++ src/components/SubmitButton.tsx
@@ -5,10 +5,6 @@ import { useFormStatus } from 'react-dom';
 export default function SubmitButton() {
   const { pending } = useFormStatus();
 
-  useEffect(() => {
-    console.log(pending);
-  }, [pending]);
-
   return (
     <button
       className="bg-gray-800 hover:bg-gray-700 disabled:bg-gray-500 text-white font-bold py-2 px-4 rounded focus:outline-none focus:shadow-outline w-32"
